cambios

+++ ms-asistencia/src/main/java/cl/instituto/pacifico/ms_asistencia/service/AsistenciaService.java
@@ -1,12 +1,15 @@
 package cl.instituto.pacifico.ms_asistencia.service;
 
 import cl.instituto.pacifico.ms_asistencia.dto.EstudianteDTO;
+import cl.instituto.pacifico.ms_asistencia.exception.BusinessException;
+import cl.instituto.pacifico.ms_asistencia.exception.ResourceNotFoundException;
 import cl.instituto.pacifico.ms_asistencia.model.Asistencia;
 import cl.instituto.pacifico.ms_asistencia.repository.AsistenciaRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 import org.springframework.web.reactive.function.client.WebClient;
+import reactor.core.publisher.Mono;
 
 import java.time.LocalDate;
 import java.util.List;
@@ -28,7 +31,7 @@ public class AsistenciaService {
             .baseUrl("http://localhost:8081")
             .defaultHeaders(headers ->
             headers.setBasicAuth("admin", "1234"))
-            .build();;
+            .build();
 
     // CREAR ASISTENCIA
     public Asistencia crear(Asistencia asistencia){
@@ -39,20 +42,24 @@ public class AsistenciaService {
         EstudianteDTO e = client.get()
                 .uri("/api/v1/estudiantes/rut/"+asistencia.getRutEstudiante()) // endpoint al que llama
                 .retrieve() // ejecuta la llamada
+                .onStatus(status -> status.is4xxClientError(),
+                        response -> {
+                            log.warn("Estudiante no existe{}", asistencia.getEstudianteId());
+                            return Mono.error(new BusinessException("El estudiante indicado no existe"));
+                        })
+                .onStatus(status -> status.is5xxServerError(),
+                        response -> {
+                            log.error("Error en ms-estudiantes");
+                            return Mono.error(new BusinessException("Error en servicio de estudiantes"));
+                        })
                 .bodyToMono(EstudianteDTO.class)// cambia de json a obj
                 .block(); // detiene el flujo hasta recibir una respuesta
-        // Validar que exista
-        if (e == null) {
-            log.error("El estudiante no existe");
-            throw new RuntimeException("Estudiante no existe");
-        }
 
         //  Copiar datos al modelo Asistencia
         asistencia.setRutEstudiante(e.getRut());
         asistencia.setEstudianteId(e.getId());
         asistencia.setNombreEstudiante(e.getNombre());
         asistencia.setFechaAsistencia(LocalDate.now());
-
 
         // GUARDAR
         log.info("Asistencia creada correctamente");
@@ -67,49 +74,61 @@ public class AsistenciaService {
 
     // BUSCAR POR ID
     public Asistencia buscarPorId(Long id) {
-        log.info("Buscando Asistencias con ID: {}", id);
-        Optional<Asistencia> asistencia = asistenciaRepository.findById(id);
-        return asistencia.orElse(null);
+        log.info("Buscando Asistencia con ID: {}", id);
+        return asistenciaRepository.findById(id).orElseThrow(() -> {
+                    log.warn("Asistencia no encontrada con ID: {}", id);
+                    return new  ResourceNotFoundException("Asistencia con id " + id + " no encontrada");
+                });
     }
 
     // OBTENER ASISTTENCIA POR RUT ESTUDIANTE
-    public List<Asistencia> obtenerPorRut(String rutEstudiante){
+    public List<Asistencia> obtenerPorRut(String rutEstudiante) {
         log.info("Buscando Asistencias con Rut: {}", rutEstudiante);
-        return asistenciaRepository.findByRutEstudiante(rutEstudiante);
-    }
 
-    // SI EXISTE X ID
-    public boolean existePorId(Long id){
-        return asistenciaRepository.existsById(id);
+        List<Asistencia> lista = asistenciaRepository.findByRutEstudiante(rutEstudiante);
+
+        if (lista.isEmpty()) {
+            throw new ResourceNotFoundException("No existen asistencias para el rut " + rutEstudiante);
+        }
+
+        return lista;
     }
 
     // ELIMINAR
     public void eliminar(Long id) {
-        log.info("Eliminando Asistencias con ID: {}", id);
+        log.info("Eliminando Asistencia con ID: {}", id);
+
+        if (!asistenciaRepository.existsById(id)) {
+            throw new ResourceNotFoundException("No se puede eliminar. Asistencia con id " + id + " no existe");
+        }
+
         asistenciaRepository.deleteById(id);
-        log.info("Asistencias eliminada correctamente");
+        log.info("Asistencia eliminada correctamente");
     }
 
     // ACTUALIZAR LOS DATOS DE LA ASISTENCIA
-    public Optional<Asistencia>actualizarCompleta(Long id, Asistencia asistenciaActualizadol){
-        return asistenciaRepository.findById(id).map(asistencia -> {
+    public Asistencia actualizarCompleta(Long id, Asistencia asistenciaActualizada) {
 
-            EstudianteDTO e = client.get()
-                    .uri("/api/v1/estudiantes/rut/"+ asistenciaActualizadol.getRutEstudiante())
-                    .retrieve()
-                    .bodyToMono(EstudianteDTO.class)
-                    .block();
+        Asistencia asistencia = asistenciaRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException(
+                        "Asistencia con id " + id + " no encontrada"
+                ));
 
-            if (e == null) {
-                throw new RuntimeException("El estudiante no existe");
-            }
+        EstudianteDTO e = client.get()
+                .uri("/api/v1/estudiantes/rut/" + asistenciaActualizada.getRutEstudiante())
+                .retrieve()
+                .bodyToMono(EstudianteDTO.class)
+                .block();
 
-            asistencia.setRutEstudiante(e.getRut());
-            asistencia.setEstudianteId(e.getId());
-            asistencia.setNombreEstudiante(e.getNombre());
+        if (e == null) {
+            log.warn("La asistencia indicada no existe");
+            throw new BusinessException("El estudiante no existe");
+        }
 
-            log.info("Asistencia actualizada correctamente");
-            return asistenciaRepository.save(asistencia);
-        });
+        asistencia.setRutEstudiante(e.getRut());
+        asistencia.setEstudianteId(e.getId());
+        asistencia.setNombreEstudiante(e.getNombre());
+
+        return asistenciaRepository.save(asistencia);
     }
 }
